+++ InvenTree/plugin/mixins/__init__.py
@@ -6,10 +6,13 @@ from common.notifications import (BulkNotificationMethod,
 from ..base.action.mixins import ActionMixin
 from ..base.barcodes.mixins import BarcodeMixin
 from ..base.event.mixins import EventMixin
-from ..base.integration.mixins import (APICallMixin, AppMixin, NavigationMixin,
-                                       PanelMixin, ScheduleMixin,
-                                       SettingsContentMixin, SettingsMixin,
-                                       UrlsMixin, ValidationMixin)
+from ..base.integration.AppMixin import AppMixin
+from ..base.integration.mixins import (APICallMixin, NavigationMixin,
+                                       PanelMixin, SettingsContentMixin,
+                                       ValidationMixin)
+from ..base.integration.ScheduleMixin import ScheduleMixin
+from ..base.integration.SettingsMixin import SettingsMixin
+from ..base.integration.UrlsMixin import UrlsMixin
 from ..base.label.mixins import LabelPrintingMixin
 from ..base.locate.mixins import LocateMixin
